Refactor
* Fix bug in test cases in which behavior of POST/PUT returning records
  was not tested
* Add spaces to aliases in test cases
* Move repeated 'get a record or None' logic into own function
* Use flask.request.get_json(force=True)
* Add statements to asserts in test code
* Small changes and fixes suggested by reviewers
* Move tests for deprecated alias endpoints to separate file
