+++ tests/test_aliases_endpoints.py
@@ -31,9 +31,9 @@ def payload_to_list(alias_payload):
     return [record["value"] for record in alias_payload["aliases"]]
 
 
-def create_random_record(client, user):
+def create_record(client, user):
     """
-    Creates a random record in indexd and returns that record's GUID.
+    Creates a record in indexd and returns that record's GUID.
     """
     document = {
         "form": "object",
@@ -54,9 +54,9 @@ def create_random_record(client, user):
 @pytest.fixture(scope="function")
 def guid(client, user):
     """
-    Creates a random record in indexd and returns that record's GUID.
+    Creates a record in indexd and returns that record's GUID.
     """
-    return create_random_record(client, user)
+    return create_record(client, user)
 
 
 @pytest.fixture(scope="function")
@@ -66,7 +66,7 @@ def aliases(client, user, guid):
     """
     aliases = [
         ".?=G}k@Up3LIlv+p96yaI06,t@?j=ejk[%+",
-        'Fa"uW<A"/\'hELmTjH%r%6@Tp^HaB^',
+        'Fa"uW< A"/\'hELmTjH%r%6@Tp^HaB^',
         "{j'8D6d5fc]5#[*9n%|G9\"hZ?z3:wX",
         '"U7XsT+EXD|1?@$ywDV"ce<B}7v9t)',
     ]
@@ -74,8 +74,7 @@ def aliases(client, user, guid):
 
     url = get_endpoint(guid)
     res = client.put(url, json=alias_payload, headers=user)
-    print("DEBUG", alias_payload)
-    assert res.status_code == 200
+    assert res.status_code == 200, f"Unable to PUT aliases: {alias_payload}"
 
     return aliases
 
@@ -105,9 +104,9 @@ def test_global_endpoint_valid_alias(client, guid, aliases):
     """
     for alias in aliases:
         res = client.get("/" + url_encode(alias))
-        assert res.status_code == 200
+        assert res.status_code == 200, res.text
         record = res.get_json()
-        assert record["did"] == guid
+        assert record["did"] == guid, f"Did not retrieve correct record for alias"
 
 
 def test_global_endpoint_nonexistant_alias(client, guid, aliases):
@@ -116,7 +115,7 @@ def test_global_endpoint_nonexistant_alias(client, guid, aliases):
     """
     fake_alias = aliases[0] + "_but_fake"
     res = client.get("/" + url_encode(fake_alias))
-    assert res.status_code == 404
+    assert res.status_code == 404, f"Request for fake alias should have failed"
 
 
 def test_global_endpoint_alias_guid_collision(client, guid, aliases, user):
@@ -126,16 +125,16 @@ def test_global_endpoint_alias_guid_collision(client, guid, aliases, user):
     """
     # create a new guid
     guid_A = guid
-    guid_B = create_random_record(client, user)
+    guid_B = create_record(client, user)
     # assign an alias with the value of guid_B to guid_A
     aliases_payload = to_payload([guid_B])
     res = client.post(get_endpoint(guid_A), json=aliases_payload, headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect that a search for value of guid_B on the global endpoint should resolve
     # to guid_B's record, not to guid_A's record
     res = client.get("/" + url_encode(guid_B))
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     record = res.get_json()
     assert record["did"] == guid_B
     assert record["did"] != guid_A
@@ -151,15 +150,17 @@ def test_global_endpoint_alias_endpoint_collision(client, guid, aliases, user):
     # on the root of the api.
     aliases_payload = to_payload(["index"])
     res = client.post(get_endpoint(guid), json=aliases_payload, headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
-    # when we query "_status", we expect to see the "_status" endpoint, not
+    # when we query "index", we expect to see the "index" endpoint, not
     # an alias.
     res = client.get("index")
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     body = res.get_json()
-    body_is_not_record = body.get("did", None) is None
-    assert body_is_not_record
+    body_is_not_record = body.get("did") is None
+    assert (
+        body_is_not_record
+    ), f"Expected not to retrieve record -- expected to retrieve index/ endpoint"
 
 
 # GET /index/{GUID}/aliases
@@ -171,7 +172,7 @@ def test_GET_aliases_invalid_GUID(client, guid, aliases):
     fake_guid = guid + "but_fake"
     alias_endpoint = get_endpoint(fake_guid)
     res = client.get(alias_endpoint)
-    assert res.status_code == 404
+    assert res.status_code == 404, res.text
 
 
 def test_GET_aliases_valid_GUID(client, guid, aliases):
@@ -180,7 +181,7 @@ def test_GET_aliases_valid_GUID(client, guid, aliases):
     """
     alias_endpoint = get_endpoint(guid)
     res = client.get(alias_endpoint)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = aliases
     assert set(aliases_in_db) == set(expected_aliases)
@@ -196,15 +197,17 @@ def test_POST_aliases_valid_GUID_valid_aliases(
     """
     new_aliases = unused_aliases
     new_alias_payload = to_payload(new_aliases)
+
+    # expect POST to return list of all aliases
     res = client.post(get_endpoint(guid), json=new_alias_payload, headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
+    aliases_in_db = payload_to_list(res.get_json())
 
     # expect new aliases to be appended to old aliases
-    res = client.get(get_endpoint(guid))
-    assert res.status_code == 200
-    aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = aliases + new_aliases
-    assert set(aliases_in_db) == set(expected_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), f"Expected to append {new_aliases} to {aliases}"
 
 
 def test_POST_aliases_unauthenticated(client, user, guid, aliases, unused_aliases):
@@ -215,14 +218,14 @@ def test_POST_aliases_unauthenticated(client, user, guid, aliases, unused_aliase
     new_aliases_payload = to_payload(new_aliases)
 
     res = client.post(get_endpoint(guid), json=new_aliases_payload)
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
     bad_user = {
         "Authorization": "Basic badpassword",
         "Content-Type": "application/json",
     }
     res = client.post(get_endpoint(guid), json=new_aliases_payload, headers=bad_user)
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
 
 def test_POST_aliases_invalid_GUID(client, user, guid, aliases, unused_aliases):
@@ -235,7 +238,7 @@ def test_POST_aliases_invalid_GUID(client, user, guid, aliases, unused_aliases):
     new_aliases_payload = to_payload(new_aliases)
 
     res = client.post(alias_endpoint, json=new_aliases_payload, headers=user)
-    assert res.status_code == 404
+    assert res.status_code == 404, res.text
 
 
 def test_POST_aliases_nonunique_aliases(client, user, guid, aliases, unused_aliases):
@@ -244,27 +247,29 @@ def test_POST_aliases_nonunique_aliases(client, user, guid, aliases, unused_alia
     already associated with another GUID
     """
     guid_A = guid
-    guid_B = create_random_record(client, user)
+    guid_B = create_record(client, user)
 
     # add unused_aliases to guid_B
     res = client.post(
         get_endpoint(guid_B), json=to_payload(unused_aliases), headers=user
     )
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect that an attempt to add unused_aliases to guid_A
     # will fail, as aliases are already assigned to guid_B
     res = client.post(
         get_endpoint(guid_A), json=to_payload(unused_aliases), headers=user
     )
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
     # expect aliases that were already associated with guid_A to be unchanged.
     res = client.get(get_endpoint(guid_A))
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = aliases
-    assert set(aliases_in_db) == set(expected_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), f"Expected original aliases {aliases} to be in db"
 
 
 def test_POST_aliases_GUID_already_has_alias(client, user, guid, aliases):
@@ -279,7 +284,7 @@ def test_POST_aliases_GUID_already_has_alias(client, user, guid, aliases):
     res = client.post(
         get_endpoint(guid), json=to_payload(subset_old_aliases), headers=user
     )
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
 
 def test_POST_aliases_duplicate_aliases_in_request(
@@ -300,14 +305,16 @@ def test_POST_aliases_duplicate_aliases_in_request(
     res = client.post(
         get_endpoint(guid), json=to_payload(duplicated_new_aliases), headers=user
     )
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
     # expect aliases in db to be unchanged
     res = client.get(get_endpoint(guid))
     aliases_in_db = payload_to_list(res.get_json())
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     expected_aliases = aliases
-    assert set(aliases_in_db) == set(expected_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), f"Expected original aliases {aliases} to be in db"
 
 
 def test_POST_aliases_valid_GUID_empty_aliases(client, user, guid, aliases):
@@ -317,14 +324,16 @@ def test_POST_aliases_valid_GUID_empty_aliases(client, user, guid, aliases):
     # POST an empty list of aliases
     empty_aliases = []
     res = client.post(get_endpoint(guid), json=to_payload(empty_aliases), headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect aliases in db to be unchanged
     res = client.get(get_endpoint(guid))
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = aliases
-    assert set(aliases_in_db) == set(expected_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), f"Expected original aliases {aliases} to be in db"
 
 
 def test_POST_no_body(client, user, guid, aliases):
@@ -332,7 +341,7 @@ def test_POST_no_body(client, user, guid, aliases):
     expect POST with no body in request to fail with 400
     """
     res = client.post(get_endpoint(guid), headers=user)
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
 
 def test_POST_bad_content_type(client, user, guid, aliases):
@@ -340,7 +349,7 @@ def test_POST_bad_content_type(client, user, guid, aliases):
     expect POST with a non-JSON content type (e.g., `text/plain`) to fail with 400
     """
     res = client.post(get_endpoint(guid), headers=user, content_type="text/plain")
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
 
 # PUT /index/{GUID}/aliases
@@ -354,15 +363,20 @@ def test_PUT_aliases_valid_GUID_valid_aliases(
     """
     new_aliases = unused_aliases
     new_alias_payload = to_payload(new_aliases)
+    # expect PUT to return list of all aliases
     res = client.put(get_endpoint(guid), json=new_alias_payload, headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
+    aliases_in_db = payload_to_list(res.get_json())
+    new_aliases = unused_aliases
 
     # expect new aliases to have replaced old aliases
     res = client.get(get_endpoint(guid))
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = new_aliases
-    assert set(aliases_in_db) == set(expected_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), f"Expect aliases in db to be {new_aliases}"
 
 
 def test_PUT_aliases_unauthenticated(client, user, guid, aliases, unused_aliases):
@@ -373,14 +387,14 @@ def test_PUT_aliases_unauthenticated(client, user, guid, aliases, unused_aliases
     new_aliases_payload = to_payload(new_aliases)
 
     res = client.put(get_endpoint(guid), json=new_aliases_payload)
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
     bad_user = {
         "Authorization": "Basic badpassword",
         "Content-Type": "application/json",
     }
     res = client.put(get_endpoint(guid), json=new_aliases_payload, headers=bad_user)
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
 
 def test_PUT_aliases_invalid_GUID(client, user, guid, aliases, unused_aliases):
@@ -393,7 +407,7 @@ def test_PUT_aliases_invalid_GUID(client, user, guid, aliases, unused_aliases):
     new_aliases_payload = to_payload(new_aliases)
 
     res = client.put(alias_endpoint, json=new_aliases_payload, headers=user)
-    assert res.status_code == 404
+    assert res.status_code == 404, res.text
 
 
 def test_PUT_aliases_nonunique_aliases(client, user, guid, aliases, unused_aliases):
@@ -404,23 +418,25 @@ def test_PUT_aliases_nonunique_aliases(client, user, guid, aliases, unused_alias
     new_aliases = unused_aliases
     # add a subset of the generated aliases to a different GUID.
     other_guid_aliases = [new_aliases[0]]
-    other_guid = create_random_record(client, user)
+    other_guid = create_record(client, user)
     res = client.put(
         get_endpoint(other_guid), json=to_payload(other_guid_aliases), headers=user
     )
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect that an attempt to add the original set of random aliases
     # will fail, as some of the aliases are already assigned to a different GUID.
     res = client.put(get_endpoint(guid), json=to_payload(new_aliases), headers=user)
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
     # expect aliases that were already associated with GUID to be unchanged.
     res = client.get(get_endpoint(guid))
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = aliases
-    assert set(aliases_in_db) == set(expected_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), f"Expected original aliases {aliases} to be in db"
 
 
 def test_PUT_aliases_previously_used_valid_alias(
@@ -434,19 +450,19 @@ def test_PUT_aliases_previously_used_valid_alias(
 
     # add a subset of the generated aliases to a different GUID
     other_guid_aliases = [new_aliases[0]]
-    other_guid = create_random_record(client, user)
+    other_guid = create_record(client, user)
     res = client.put(
         get_endpoint(other_guid), json=to_payload(other_guid_aliases), headers=user
     )
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # delete the aliases from the different GUID
     res = client.delete(get_endpoint(other_guid), headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect an attempt to add the original set of random aliases will succeed
     res = client.put(get_endpoint(guid), json=to_payload(new_aliases), headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
 
 def test_PUT_aliases_GUID_already_has_alias(client, user, guid, aliases):
@@ -461,14 +477,16 @@ def test_PUT_aliases_GUID_already_has_alias(client, user, guid, aliases):
     res = client.put(
         get_endpoint(guid), json=to_payload(subset_old_aliases), headers=user
     )
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect the aliases of this GUID to be the new subset of the old aliases
     res = client.get(get_endpoint(guid))
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = subset_old_aliases
-    assert set(aliases_in_db) == set(subset_old_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), f"Expected aliases {expected_aliases} to be in db"
 
 
 def test_PUT_aliases_duplicate_aliases_in_request(
@@ -490,7 +508,7 @@ def test_PUT_aliases_duplicate_aliases_in_request(
     res = client.put(
         get_endpoint(guid), json=to_payload(duplicated_new_aliases), headers=user
     )
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
 
 def test_PUT_aliases_valid_GUID_empty_aliases(client, user, guid, aliases):
@@ -499,14 +517,16 @@ def test_PUT_aliases_valid_GUID_empty_aliases(client, user, guid, aliases):
     """
     empty_aliases = []
     res = client.put(get_endpoint(guid), json=to_payload(empty_aliases), headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect aliases in db to be empty
     res = client.get(get_endpoint(guid))
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = empty_aliases
-    assert set(aliases_in_db) == set(expected_aliases)
+    assert set(aliases_in_db) == set(
+        expected_aliases
+    ), "Expected no aliases to be in db"
 
 
 def test_PUT_no_body(client, user, guid, aliases):
@@ -514,7 +534,7 @@ def test_PUT_no_body(client, user, guid, aliases):
     expect PUT with no body in request to fail with 400
     """
     res = client.put(get_endpoint(guid), json=None, headers=user)
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
 
 def test_PUT_bad_content_type(client, user, guid, aliases):
@@ -522,7 +542,7 @@ def test_PUT_bad_content_type(client, user, guid, aliases):
     expect PUT with a non-JSON content type (e.g., `text/plain`) to fail with 400
     """
     res = client.put(get_endpoint(guid), headers=user, content_type="text/plain")
-    assert res.status_code == 400
+    assert res.status_code == 400, res.text
 
 
 # DELETE /index/{GUID}/aliases
@@ -532,13 +552,13 @@ def test_DELETE_all_aliases_valid_GUID(client, user, guid, aliases):
     normal operation: expect to delete all aliases if valid GUID 
     """
     res = client.delete(get_endpoint(guid), headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect all aliases to be gone
     res = client.get(get_endpoint(guid))
     aliases_in_db = payload_to_list(res.get_json())
     expected_aliases = []
-    assert aliases_in_db == expected_aliases
+    assert aliases_in_db == expected_aliases, "Expected no aliases to be in db"
 
 
 def test_DELETE_all_aliases_unauthenticated(client, user, guid, aliases):
@@ -546,14 +566,14 @@ def test_DELETE_all_aliases_unauthenticated(client, user, guid, aliases):
     expect request to fail with 403 if user is unauthenticated
     """
     res = client.delete(get_endpoint(guid))
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
     bad_user = {
         "Authorization": "Basic badpassword",
         "Content-Type": "application/json",
     }
     res = client.delete(get_endpoint(guid), headers=bad_user)
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
 
 def test_DELETE_all_aliases_invalid_GUID(client, user, guid, aliases):
@@ -562,7 +582,7 @@ def test_DELETE_all_aliases_invalid_GUID(client, user, guid, aliases):
     """
     fake_guid = guid + "_but_fake"
     res = client.delete(get_endpoint(fake_guid), headers=user)
-    assert res.status_code == 404
+    assert res.status_code == 404, res.text
 
 
 # DELETE /index/{GUID}/aliases/{ALIAS}
@@ -576,7 +596,7 @@ def test_DELETE_one_alias_valid_GUID(client, user, guid, aliases):
     alias_to_delete = aliases[0]
     endpoint = get_endpoint(guid) + "/" + url_encode(alias_to_delete)
     res = client.delete(endpoint, headers=user)
-    assert res.status_code == 200
+    assert res.status_code == 200, res.text
 
     # expect that alias to no longer be in this guid's aliases
     res = client.get(get_endpoint(guid))
@@ -595,14 +615,14 @@ def test_DELETE_one_alias_unauthenticated(client, user, guid, aliases):
 
     # expect an unauthenticated delete request to that alias to fail
     res = client.delete(endpoint)
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
     bad_user = {
         "Authorization": "Basic badpassword",
         "Content-Type": "application/json",
     }
     res = client.delete(endpoint, headers=bad_user)
-    assert res.status_code == 403
+    assert res.status_code == 403, res.text
 
 
 def test_DELETE_one_alias_invalid_GUID(client, user, guid, aliases):
@@ -616,7 +636,7 @@ def test_DELETE_one_alias_invalid_GUID(client, user, guid, aliases):
     alias_to_delete = "fake_alias"
     endpoint = get_endpoint(fake_guid) + "/" + url_encode(alias_to_delete)
     res = client.delete(endpoint, headers=user)
-    assert res.status_code == 404
+    assert res.status_code == 404, res.text
 
 
 def test_DELETE_one_alias_GUID_does_not_have_alias(client, user, guid, aliases):
@@ -628,4 +648,4 @@ def test_DELETE_one_alias_GUID_does_not_have_alias(client, user, guid, aliases):
     alias_to_delete = "fake_alias"
     endpoint = get_endpoint(guid) + "/" + url_encode(alias_to_delete)
     res = client.delete(endpoint, headers=user)
-    assert res.status_code == 404
+    assert res.status_code == 404, res.text
